message

+++ ft_strlcat.c
@@ -6,7 +6,7 @@
 /*   By: tnanchen <thomasnanchen@hotmail.com>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/14 16:14:15 by tnanchen          #+#    #+#             */
-/*   Updated: 2021/10/15 18:39:01 by tnanchen         ###   ########.fr       */
+/*   Updated: 2021/10/19 20:07:21 by tnanchen         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,11 +14,11 @@
 
 size_t	ft_strlcat(char *dst, const char *src, size_t dstsize)
 {
-	int		i;
-	int		n;
+	size_t	i;
+	size_t	n;
 	size_t	src_length;
 	size_t	dst_length;
-	int		src_length_cpy;
+	size_t	src_length_cpy;
 
 	src_length = 0;
 	while (src[src_length])
